+++ thonny/plugins/rpi_pico/__init__.py
@@ -8,24 +8,16 @@ from thonny.plugins.micropython import (
     BareMetalMicroPythonConfigPage,
 )
 from thonny.plugins.micropython.uf2dialog import Uf2FlashingDialog
+from thonny.plugins.rp2040 import RP2040BackendProxy, RP2040BackendConfigPage
 from thonny.ui_utils import show_dialog
 
 logger = getLogger(__name__)
 
 
-class RaspberryPiPicoBackendProxy(BareMetalMicroPythonProxy):
-    @property
-    def consider_unknown_devices(self):
-        return False
-
+class RaspberryPiPicoBackendProxy(RP2040BackendProxy):
     @classmethod
     def should_consider_unknown_devices(cls):
         return False
-
-    @property
-    def known_usb_vids_pids(self):
-        # Required for backward compatibility
-        return RaspberryPiPicoBackendProxy.get_known_usb_vids_pids()
 
     @classmethod
     def get_known_usb_vids_pids(cls):
@@ -34,7 +26,7 @@ class RaspberryPiPicoBackendProxy(BareMetalMicroPythonProxy):
 
     @classmethod
     def device_is_present_in_bootloader_mode(cls):
-        return bool(PicoFlashingDialog.get_possible_targets())
+        return bool(Uf2FlashingDialog.get_possible_targets())
 
     def get_node_label(self):
         return "Raspberry Pi Pico"
@@ -44,13 +36,8 @@ class RaspberryPiPicoBackendProxy(BareMetalMicroPythonProxy):
         show_dialog(dlg)
         return dlg.success
 
-    def _get_backend_launcher_path(self) -> str:
-        import thonny.plugins.rpi_pico.rpi_pico_backend
 
-        return thonny.plugins.rpi_pico.rpi_pico_backend.__file__
-
-
-class RaspberryPiPicoBackendConfigPage(BareMetalMicroPythonConfigPage):
+class RaspberryPiPicoBackendConfigPage(RP2040BackendConfigPage):
     def _has_flashing_dialog(self):
         return True
 
@@ -121,7 +108,14 @@ def load_plugin():
     # The main reason is to reduce the number of items in the backend switcher menu
     import thonny.plugins.circuitpython
     import thonny.plugins.micropython
+    import thonny.plugins.esp
 
+    thonny.plugins.circuitpython.VIDS_PIDS_TO_AVOID.update(
+        RaspberryPiPicoBackendProxy.get_known_usb_vids_pids()
+    )
     thonny.plugins.micropython.VIDS_PIDS_TO_AVOID_IN_GENERIC_BACKEND.update(
+        RaspberryPiPicoBackendProxy.get_known_usb_vids_pids()
+    )
+    thonny.plugins.esp.VIDS_PIDS_TO_AVOID_IN_ESP_BACKENDS.update(
         RaspberryPiPicoBackendProxy.get_known_usb_vids_pids()
     )